refactor : 파일 디렉토리 구조 벽녕 (#559)
