+++ src/main/java/com/clubber/ClubberServer/domain/admin/dto/UpdateClubPageResponse.java
@@ -2,7 +2,7 @@ package com.clubber.ClubberServer.domain.admin.dto;
 
 import com.clubber.ClubberServer.domain.club.domain.Club;
 import com.clubber.ClubberServer.domain.club.domain.ClubInfo;
-import com.clubber.ClubberServer.global.vo.ImageVO;
+import com.clubber.ClubberServer.global.vo.image.ImageVO;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
