+++ src/main/java/com/clubber/ClubberServer/domain/club/domain/Department.java
@@ -1,7 +1,7 @@
 package com.clubber.ClubberServer.domain.club.domain;
 
 
-import com.clubber.ClubberServer.global.enummapper.EnumMapperType;
+import com.clubber.ClubberServer.global.mapper.enums.EnumMapperType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
