+++ src/main/java/com/clubber/ClubberServer/domain/notice/domain/Notice.java
@@ -2,7 +2,7 @@ package com.clubber.ClubberServer.domain.notice.domain;
 
 import com.clubber.ClubberServer.domain.common.BaseEntity;
 import com.clubber.ClubberServer.domain.owner.domain.Owner;
-import com.clubber.ClubberServer.global.vo.ImageVO;
+import com.clubber.ClubberServer.global.vo.image.ImageVO;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
