+++ src/main/java/com/clubber/ClubberServer/domain/recruit/dto/GetOneRecruitInListResponse.java
@@ -1,14 +1,11 @@
 package com.clubber.ClubberServer.domain.recruit.dto;
 
 import com.clubber.ClubberServer.domain.recruit.domain.Recruit;
-import com.clubber.ClubberServer.global.vo.ImageVO;
-import com.fasterxml.jackson.annotation.JsonFormat;
+import com.clubber.ClubberServer.global.vo.image.ImageVO;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-
-import java.time.LocalDateTime;
 
 @Getter
 @Builder(access = AccessLevel.PRIVATE)
