+++ src/main/java/com/clubber/ClubberServer/global/converter/EnumMapperTypeConverter.java
@@ -1,7 +1,9 @@
-package com.clubber.ClubberServer.global.enummapper;
+package com.clubber.ClubberServer.global.converter;
 
 import com.clubber.ClubberServer.global.exception.EnumTypeNotValidException;
 
+import com.clubber.ClubberServer.global.mapper.enums.EnumMapper;
+import com.clubber.ClubberServer.global.mapper.enums.EnumMapperType;
 import jakarta.persistence.AttributeConverter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
