+++ src/main/java/com/clubber/ClubberServer/global/vo/enums/EnumMapperVO.java
@@ -1,5 +1,6 @@
-package com.clubber.ClubberServer.global.enummapper;
+package com.clubber.ClubberServer.global.vo.enums;
 
+import com.clubber.ClubberServer.global.mapper.enums.EnumMapperType;
 import lombok.Getter;
 
 @Getter
