+++ src/test/java/com/clubber/ClubberServer/global/util/ImageUtilTest.java
@@ -1,9 +1,9 @@
-package com.clubber.ClubberServer.global.image;
+package com.clubber.ClubberServer.global.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static com.clubber.ClubberServer.global.jwt.JwtStatic.IMAGE_SERVER;
+import static com.clubber.ClubberServer.global.common.ClubberStatic.IMAGE_SERVER;
 import static org.assertj.core.api.Assertions.*;
 public class ImageUtilTest {
     private final String exampleKey = "imagekey";
